Extract TestContext into external source file.
