+++ tools/slang-test/slang-test.vcxproj
@@ -164,11 +164,13 @@
   <ItemGroup>
     <ClInclude Include="os.h" />
     <ClInclude Include="render-api-util.h" />
+    <ClInclude Include="test-context.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp" />
     <ClCompile Include="os.cpp" />
     <ClCompile Include="render-api-util.cpp" />
+    <ClCompile Include="test-context.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\source\core\core.vcxproj">
